@@ -46,67 +42,71 @@
   },
   "homepage": "https://github.com/richardcyrus/fm-rmx-product-feedback-app#readme",
   "dependencies": {
-    "@prisma/client": "^4.10.1",
-    "@radix-ui/react-label": "^2.0.0",
-    "@radix-ui/react-select": "^1.2.0",
-    "@radix-ui/react-tabs": "^1.0.2",
-    "@remix-run/node": "1.13",
-    "@remix-run/react": "1.13",
-    "@remix-run/vercel": "1.13",
-    "@vercel/node": "^2.9.8",
+    "@prisma/client": "^5.8.0",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@remix-run/css-bundle": "2.5.1",
+    "@remix-run/node": "2.5.1",
+    "@remix-run/react": "2.5.1",
+    "@remix-run/serve": "2.5.1",
+    "isbot": "^4.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tiny-invariant": "^1.3.1",
-    "zod": "^3.20.6"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@faker-js/faker": "^7.6.0",
-    "@remix-run/dev": "1.13",
-    "@remix-run/eslint-config": "1.13",
-    "@remix-run/serve": "1.13",
-    "@svgr/cli": "^6.5.1",
-    "@svgr/plugin-jsx": "^6.5.1",
-    "@svgr/plugin-prettier": "^6.5.1",
-    "@svgr/plugin-svgo": "^6.5.1",
-    "@testing-library/cypress": "^9.0.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/eslint": "^8.21.1",
-    "@types/node": "^18.14.2",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@vitejs/plugin-react": "^3.1.0",
-    "autoprefixer": "^10.4.13",
-    "axe-core": "^4.6.3",
-    "concurrently": "^7.6.0",
-    "cssnano": "^5.1.15",
-    "cypress": "^12.6.0",
-    "cypress-axe": "^1.4.0",
-    "dotenv-cli": "^7.0.0",
-    "esbuild-register": "^3.4.2",
-    "eslint": "^8.35.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "happy-dom": "^8.9.0",
+    "@faker-js/faker": "^8.3.1",
+    "@remix-run/dev": "2.5.1",
+    "@svgr/cli": "^8.1.0",
+    "@svgr/plugin-jsx": "^8.1.0",
+    "@svgr/plugin-prettier": "^8.1.0",
+    "@svgr/plugin-svgo": "^8.1.0",
+    "@testing-library/cypress": "^10.0.1",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/eslint": "^8.56.2",
+    "@types/node": "^20.11.0",
+    "@types/react": "^18.2.47",
+    "@types/react-dom": "^18.2.18",
+    "@typescript-eslint/eslint-plugin": "^6.7.4",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.16",
+    "axe-core": "^4.8.3",
+    "concurrently": "^8.2.2",
+    "cssnano": "^6.0.3",
+    "cypress": "^13.6.2",
+    "cypress-axe": "^1.5.0",
+    "dotenv-cli": "^7.3.0",
+    "esbuild-register": "^3.5.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "happy-dom": "^13.0.6",
     "husky": "^8.0.3",
-    "lint-staged": "^13.1.2",
+    "lint-staged": "^15.2.0",
     "micromatch": "^4.0.5",
-    "postcss": "^8.4.21",
-    "postcss-cli": "^10.1.0",
-    "postcss-import": "^15.1.0",
+    "postcss": "^8.4.33",
+    "postcss-cli": "^11.0.0",
+    "postcss-import": "^16.0.0",
     "postcss-nested": "^6.0.1",
-    "postcss-sort-media-queries": "^4.3.0",
-    "prettier": "^2.8.4",
-    "prisma": "^4.10.1",
-    "start-server-and-test": "^1.15.4",
-    "stylelint": "^15.2.0",
-    "stylelint-config-standard": "^30.0.1",
-    "stylelint-order": "^6.0.2",
-    "typescript": "^4.9.5",
-    "vite-tsconfig-paths": "^4.0.5",
-    "vitest": "^0.29.1"
+    "postcss-sort-media-queries": "^5.2.0",
+    "prettier": "^3.2.2",
+    "prisma": "^5.8.0",
+    "start-server-and-test": "^2.0.3",
+    "stylelint": "^16.1.0",
+    "stylelint-config-standard": "^36.0.0",
+    "stylelint-order": "^6.0.4",
+    "typescript": "^5.1.6",
+    "vite-tsconfig-paths": "^4.2.3",
+    "vitest": "^1.2.0"
   },
   "browserslist": {
     "production": [
@@ -122,10 +122,10 @@
     ]
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=18.19.0"
   },
   "sideEffects": false,
   "volta": {
-    "node": "18.14.2"
+    "node": "18.19.0"
   }
 }
